Adjusted server configuration to use relative paths
All server configuration paths in conf/serv are now relative to an
environment variable $server_home, which is set in the init script. The
init script will resolve the realpath of this loclation, removing
symlinks and . or .. references.

TODO: Database references are still absolute and need to be adjusted to
be relative to $server_home. The function void set_path(...) in serv.c
should be moved to generic.[ch] and renamed to something more
appropriate.