Fixed header parsing for CGI/Kraknet
The internal CGI header Status is no longer passed to the client. Its
function, changing the HTTP response code and message, is still intact
according to the standard.

Empty strings will also no longer be printed in the header. Previously,
this would have made it possible to accidentally end the header section
of the response.