Adding directory listing script and link
A Bash script for listing the current directory contents in an HTML
formatted presentation has been added at bin/list.bash. Because this
implementation is likely to be replaced by a more permanent solution
with a flexible template for the list page style, a symlink to that
script at bin/list has also been created and will be used as the
canonical reference to the script.