Directory listing now available.
New feature:
When a directory is requested (URI) but contains no default document (as
specified in conf/serv) the listing of the files, generated by bin/list
is returned instead.

Bug fix:
A directory requested without the trailing slash now 301 redirects to
the proper location, including the trailing slash, and not to the name
of the default document.