Fixed non-standard code.
A while loop in src/c/kraknet.c was improperly formatted according to
the coding standard for the project. It featured curly braces for a
single statement, and had a line of code intended to run after the
execution of the loop on the same line as the closing brace, tabbed to
the same level as the internal of the loop.

The fix was to remove the braces and reduce the indentation of that
command by one level.