Temp directory and build script
The creation of a temp directory was added to the default init script.
This is needed for caching POST data as it's submitted so that it can be
passed to child processes atomically. A build script (/build.sh) was
added to expedite the initial setup for new users. It checks basic
dependencies and builds the server and related binaries cleanly.