Root check for accessing /tmp/
Although the server needs to access /tmp/ to create a working area for
POST data, it is unlikely to be able to create a folder and chmod it
unless the init script is run as root. Thus, a check is performed and
this step is skipped if you are not root.