Fixed issue "Crash if variable not defined"
Fixed an issue where get_conf_line_s was apparently crashing if it
didn't find the variable name it was looking for inside of a config
file. The actual issue was that it called unquote_str() with a NULL
pointer if the variable name was not found.

unquote_str() is now protected against NULL pointers and strings which
are too short to possibly be in quotation marks (less than 2
characters).