Logs are now chowned if possible.
If the server encounters an error after being run as root for the first
time it was possible that logs and the log directory would be owned by
root and inaccessible to the intended user which the server should run
as. These files are now chowned to the appropriate user if possible at
startup.