serv.c Handles POST Data
The server code handles munching POST data off of the client stream now,
instead of allowing it to pass to the http11 module. This removes the
need to read that data off of the main server's client stream uselessly
before resuming HTTP request handling in the event of Connection:
keep-alive mode.