Add REQUEST_URI and set DOCUMENT_ROOT
A new environment variable, REQUEST_URI is now set for each request. It
reflects the URI the request is looking for before any processing of the
QUERY_STRING.

DOCUMENT_ROOT was not being set when web_root was changed as a result of
host-based site service. The server was not impacted because interally
web_root is used everywhere. CGI scripts might expect to have
DOCUMENT_ROOT set appropriately, as it would be on Apache.